+++ changelog.md
@@ -3,6 +3,7 @@
 ## 2.2.0
 
 - 支持eureka注册中心，需要执行`sop-2.2.0.sql`升级文件
+- 签名内容支持urlencode（设置`sign.urlencode=true`）
 
 ## 2.1.3